Improve handling of the tally module.
The pam_tally module, only provided by Linux-PAM for us, is replaced in
newer instances with the pam_tally2 module which is wordsize-independent.

For this reason, make the configuration choose the best tally module
between the two depending on the version of the Linux-PAM package.

Also drop the DEBUG indication: the module does not support the debug
option.